+++ libarchive/archive_read_support_filter_zstd.c
@@ -116,10 +116,10 @@ zstd_bidder_bid(struct archive_read_filter_bidder *self,
 	 * Regular frames contain a 4 byte magic number followed by a 2-14
 	 * byte frame header, some data, and a 3 byte end marker.
 	 */
-	const ssize_t min_zstd_frame_size = 8;
+	const size_t min_zstd_frame_size = 8;
 
-	ssize_t offset_in_buffer = 0;
-	const ssize_t max_lookahead = 64 * 1024;
+	size_t offset_in_buffer = 0;
+	const size_t max_lookahead = 64 * 1024;
 	uint32_t magic_number;
 
 	/* Zstd regular frame magic number. */
@@ -150,7 +150,7 @@ zstd_bidder_bid(struct archive_read_filter_bidder *self,
 		offset_in_buffer += 4;
 
 		/* Ensure that we can read another 4 bytes. */
-		if (offset_in_buffer + 4 > avail) {
+		if (offset_in_buffer + 4 > (size_t)avail) {
 			buffer = __archive_read_filter_ahead(filter,
 			    offset_in_buffer + 4, &avail);
 			if (buffer == NULL)
@@ -169,7 +169,7 @@ zstd_bidder_bid(struct archive_read_filter_bidder *self,
 		 * There should be at least one more frame
 		 * if this is zstd data.
 		 */
-		if (offset_in_buffer + min_zstd_frame_size > avail) {
+		if (offset_in_buffer + min_zstd_frame_size > (size_t)avail) {
 			if (offset_in_buffer + min_zstd_frame_size >
 			    max_lookahead)
 				return (0);
